requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0
python-dotenv
gunicorn
